Clean up
Clean up config plugins
